增加XMLDictionary 、 优化代码， 增加infer和oclint脚本

+++ MAREX/Foundation/NSFileManager+MAREX.m
@@ -314,7 +314,7 @@
         NSDictionary *fileAttributes = [[NSFileManager defaultManager] attributesOfItemAtPath:[folderPath stringByAppendingPathComponent:file] error:nil];
         folderSize += [[fileAttributes objectForKey:NSFileSize] intValue];
     }
-    return folderSize;
+    return (NSInteger)folderSize;
 }
 
 @end
